Changed more names for consistancy
