+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/src/DrivePrograms/Autonomous/old/RedAutonomousNearWarehouse.java
@@ -1,23 +1,22 @@
-package org.firstinspires.ftc.teamcode.src.DrivePrograms.Autonomous;
+package org.firstinspires.ftc.teamcode.src.DrivePrograms.Autonomous.old;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 
 import org.firstinspires.ftc.teamcode.src.Utills.AutonomousTemplate;
 
-@Autonomous(name = "BlueAutonomousNearWarehouse")
+@Autonomous(name = "RedAutonomousNearWarehouse")
 @Disabled
-public class BlueAutonomousNearWarehouse extends AutonomousTemplate {
+public class RedAutonomousNearWarehouse extends AutonomousTemplate {
 
     @Override
     public void runOpMode() throws InterruptedException {
         this.initAll();
-        odometry.setPosition(133.5, 64, 180);
         telemetry.addData("Initialization Status: ", "Complete");
         telemetry.update();
         waitForStart();
-
-        driveSystem.moveToPosition(106, 65, 2, true);
+        odometry.setPosition(7.5, 64, 180);
+        driveSystem.moveToPosition(29, 64, 1, true);
         podServos.raise();
         driveSystem.strafeAtAngle(180, 1);
         Thread.sleep(50);
